requirements.txt
aiogram==3.12.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
